file: 지워지는 데이터 기준으로 파일 위치 이동
